+++ components/Actions/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { List, Typography } from 'antd/lib';
-import { EducationTitle } from 'components/Education';
+import { EducationTitle } from 'components/Home/MintNft/Education';
 import data from './data.json';
 
 const { Text, Paragraph } = Typography;
