+++ components/Home/MintNft/Education/data.json
