+++ components/Home/MintNft/Education/index.jsx
